engine/requirements.txt
pymongo==4.6.0
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
apify-client==1.7.1
PyYAML==6.0.1